# meta/definitions/openapi.yaml
...
openapi: 3.1.0
info:
  title: TCGdex API
  description: |
    A Multilanguage Pokémon TCG Database with Cards Pictures and most of the
    informations contained on the cards.

    Error responses follow RFC 9457 standard for problem details in HTTP APIs,
    with Content-Type application/problem+json. All errors include standardized
    attributes like type, title, status, detail, endpoint, and method.

    Most endpoints that return arrays support filtering, sorting, and pagination features.
    See https://tcgdex.dev/rest/filtering-sorting-pagination for more details.

    You can find out more about TCGdex at
    [https://www.tcgdex.net](https://www.tcgdex.net) or on
    [Discord](https://discord.gg/NehYTAhsZE).
  contact:
    name: TCGdex
    url: https://github.com/tcgdex/cards-database
    email: contact@tcgdex.net
  license:
    name: MIT License
    url: https://github.com/tcgdex/cards-database/blob/master/LICENSE
  version: "2"
...
servers:
  - url: https://api.tcgdex.net/v2/{lang}
    variables:
      lang:
        enum:
          - en
          - fr
          - es
          - es-mx
          - it
          - pt
          - pt-br
          - pt-pt
          - zh-cn
          - zh-tw
          - ja
          - ko
          - th
        default: en
...
paths:
...
  /cards/search:
    get:
      tags:
        - cards
      summary: Search the card catalog
      description: Searches localized card names and rules text, with canonical gameplay and collection filters. Repeating a filter or separating values with commas matches any value within that filter; different filters are combined with AND.
      operationId: searchCardCatalog
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
          description: Accent-insensitive search across localized and English names, attacks, abilities, effects, descriptions, IDs, and illustrators.
        - name: set
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: booster
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: category
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: type
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: rarity
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: stage
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: trainerType
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: suffix
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: hasAbility
          in: query
          schema: { type: boolean }
        - name: hpMin
          in: query
          schema: { type: number, minimum: 0 }
        - name: hpMax
          in: query
          schema: { type: number, minimum: 0 }
        - name: retreat
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: weakness
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: attackCostType
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: string } }
        - name: damageMin
          in: query
          schema: { type: number, minimum: 0 }
        - name: damageMax
          in: query
          schema: { type: number, minimum: 0 }
        - name: illustrator
          in: query
          schema: { type: string, maxLength: 120 }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 48 }
      responses:
        '200':
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardCatalogSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /cards/{cardId}:
    get:
      tags:
        - cards
      summary: Find card by ID
      description: Returns detailed information about a specified card
      operationId: getCardById
      parameters:
        - name: cardId
          in: path
          description: ID of card to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://tcgdex.dev/errors/not-found"
            title: "Resource Not Found"
            status: 404
            detail: "The requested resource could not be found"
            endpoint: "/v2/cards/invalid-id"
            method: "GET"
    BadRequest:
      description: The request was invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://tcgdex.dev/errors/bad-request"
            title: "Invalid Request"
            status: 400
            detail: "The request contains invalid parameters"
            endpoint: "/v2/sets/invalid-format"
            method: "GET"
    InternalServerError:
      description: An unexpected error occurred on the server
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://tcgdex.dev/errors/server-error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred while processing your request"
            endpoint: "/v2/cards"
            method: "GET"
  schemas:
    Problem:
      description: RFC 9457 problem details object for API errors
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference that identifies the problem type
          example: "https://tcgdex.dev/errors/not-found"
        title:
          type: string
          description: Short, human-readable summary of the problem type
          example: "The resource you are trying to reach does not exist"
        status:
          type: integer
          description: HTTP status code
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence of the problem
          example: "The requested card with ID 'invalid-id' could not be found"
        endpoint:
          type: string
          description: The API endpoint that was requested
          example: "/v2/i-do-not-exists"
        method:
          type: string
          description: The HTTP method used for the request
          example: "GET"
      examples:
        notFound:
          value:
            type: "https://tcgdex.dev/errors/not-found"
            title: "Resource Not Found"
            status: 404
            detail: "The requested card with ID 'base1-999' could not be found"
            endpoint: "/v2/cards/base1-999"
            method: "GET"
        badRequest:
          value:
            type: "https://tcgdex.dev/errors/bad-request"
            title: "Invalid Request"
            status: 400
            detail: "The provided parameter 'hp' must be a valid number"
            endpoint: "/v2/hp/abc"
            method: "GET"
        serverError:
          value:
            type: "https://tcgdex.dev/errors/server-error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred while processing your request"
            endpoint: "/v2/cards"
            method: "GET"
    WeakRes:
      type: array
      xml:
        name: WeakResItem
        wrapped: true
      items:
        required:
          - type
        type: object
        properties:
          type:
            type: string
            example: "Psychic"
          value:
            type: [string, null]
            example: x2
    Booster:
      description: Represents a booster pack that contains Pokemon cards
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
          description: The booster ID using format boo_<set_id>-<booster_name>
          pattern: "^boo_.*-.*$"
          example: "boo_A1-mewtwo"
        name:
          type: string
          description: The name of the booster
          example: "Mewtwo"
        logo:
          type: [string, null]
          description: URL to the logo of the booster
          example: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-logo"
        artwork_front:
          type: [string, null]
          description: URL to the front artwork of the booster pack
          example: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-front"
        artwork_back:
          type: [string, null]
          description: URL to the back artwork of the booster pack
          example: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-back"
      examples:
        complete:
          value:
            id: "boo_A1-mewtwo"
            name: "Mewtwo"
            logo: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-logo"
            artwork_front: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-front"
            artwork_back: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-front"
        minimal:
          value:
            id: "boo_base1-base"
            name: "Base Set Booster"
            logo: null
            artwork_front: "https://assets.tcgdex.net/en/base/base1/boosters/base/front"
            artwork_back: null
...
    SetResume:
      description: Basic information about a Pokemon TCG set
      required:
        - cardCount
        - id
        - name
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the set
          example: base1
        name:
          type: string
          description: Name of the set
          example: Base Set
        logo:
          type: string
          description: URL to the logo of the set
          example: https://assets.tcgdex.net/en/base/base1/logo
        symbol:
          type: string
          description: URL to the set symbol
          example: https://assets.tcgdex.net/univ/base/base1/symbol
        cardCount:
          required:
            - official
            - total
          type: object
          description: Statistics about the number of cards in this set
          properties:
            total:
              type: number
              description: Total number of cards in the set including variants
              example: 101
            official:
              type: number
              description: Number of cards officially numbered in the set
              example: 100
    CardResume:
      description: Basic information about a Pokemon TCG card
      required:
        - id
        - localId
        - name
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the card
          example: "base1-1"
        localId:
          type: string
          description: Card number within its set
          example: "1"
        image:
          type: [string, null]
          description: URL to the card image
          example: "https://assets.tcgdex.net/en/base/base1/1"
        name:
          type: string
          description: Name of the card (including any suffix)
          example: "Alakazam"
...
    CardCatalogValue:
      type: object
      required: [value, label]
      properties:
        value: { type: string }
        label: { type: string }
...
    CardCatalogItem:
      allOf:
        - $ref: '#/components/schemas/CardResume'
        - type: object
          required: [set, boosters, category, types, hasAbility]
          properties:
            set:
              type: object
              required: [id, name]
              properties:
                id: { type: string }
                name: { type: string }
            boosters:
              type: array
              items:
                type: object
                required: [id, name]
                properties:
                  id: { type: string }
                  name: { type: string }
            category:
              $ref: '#/components/schemas/CardCatalogValue'
            types:
              type: array
              items: { $ref: '#/components/schemas/CardCatalogValue' }
            rarity:
              anyOf:
                - $ref: '#/components/schemas/CardCatalogValue'
                - { type: 'null' }
            stage:
              anyOf:
                - $ref: '#/components/schemas/CardCatalogValue'
                - { type: 'null' }
            trainerType:
              anyOf:
                - $ref: '#/components/schemas/CardCatalogValue'
                - { type: 'null' }
            suffix:
              anyOf:
                - $ref: '#/components/schemas/CardCatalogValue'
                - { type: 'null' }
            hp: { type: [number, null] }
            retreat: { type: [number, null] }
            illustrator: { type: [string, null] }
            hasAbility: { type: boolean }
    CardCatalogSearchResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CardCatalogItem'
        pagination:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page: { type: integer, minimum: 1 }
            pageSize: { type: integer, minimum: 1, maximum: 100 }
            total: { type: integer, minimum: 0 }
            totalPages: { type: integer, minimum: 1 }
...
    Card:
      description: Detailed information about a Pokemon TCG card, including game mechanics
      required:
        - category
        - id
        - localId
        - name
        - rarity
        - set
        - legal
        - updated
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the card
          example: base1-1
        localId:
          type: string
          description: Card number within its set
          example: '1'
        image:
          type: string
          description: URL to the card image
          example: https://assets.tcgdex.net/en/base/base1/1
        name:
          type: string
          description: Name of the card (including any suffix)
          example: Alakazam
        illustrator:
          type: [string, null]
          description: Artist who illustrated the card
          example: "Ken Sugimori"
        category:
          type: string
          description: Card category (Pokemon, Trainer, or Energy)
          example: "Pokemon"
        rarity:
          type: string
          description: Card rarity (Common, Uncommon, Rare, etc.)
          example: "Rare"
        set:
          $ref: '#/components/schemas/SetResume'
          description: The set this card belongs to
        variants:
          required:
            - firstEdition
            - holo
            - normal
            - reverse
            - wPromo
          type: object
          description: Indicates which variants of this card exist (overrides set variants)
          properties:
            normal:
              type: boolean
              description: Indicates whether a normal variant exists
            reverse:
              type: boolean
              description: Indicates whether a reverse holo variant exists
            holo:
              type: boolean
              description: Indicates whether a holo variant exists
            firstEdition:
              type: boolean
              description: Indicates whether a first edition variant exists
            wPromo:
              type: boolean
              description: Indicates whether a promotional variant exists
        variant_detailed:
            type: [array,"null"]
            description: Detailed information about a card variant
            items:
                type: object
                required:
                    - type
                properties:
                    type:
                        type: string
                        description: The type of variant (e.g., normal, reverse, holo, etc.)
                    size:
                        type: string
                        description: The size of the variant (e.g., standard, jumbo, etc.)
                        nullable: true
                    stamp:
                        type: array
                        description: The stamps of the variant (e.g., 'Staff', 'Pokemon-Centre', etc.)
                        items:
                            type: string
                        nullable: true
                    foil:
                        type: string
                        description: The foil of the variant (e.g., 'Pokeball', 'MasterBall', etc.)
                        nullable: true
        hp:
          type: [number, "null"]
          description: Hit Points (HP) of the Pokemon
          example: 80
        types:
          type: [array, null]
          description: Energy types of the Pokemon
          items:
            type: string
            example: Psychic
        evolveFrom:
          type: [string, null]
          description: Name of the Pokemon this evolves from
          example: "Kadabra"
        stage:
          type: [string, null]
          description: Evolution stage (Basic, Stage 1, Stage 2, etc.)
          example: "Stage 2"
        suffix:
          type: [string, null]
          description: Special card suffix (EX, GX, V, etc.)
        item:
          required:
            - effect
            - name
          type: object
          description: Pokemon's held item information
          properties:
            name:
              type: string
              description: Name of the held item
            effect:
              type: string
              description: Effect of the held item
        trainerType:
          type: [string, null]
          description: Type of trainer card (Item, Supporter, Stadium, Tool)
        energyType:
          type: [string, null]
          description: Type of energy card (Basic, Special)
        regulationMark:
          type: [string, null]
          description: Regulation mark on cards (introduced in Sword & Shield)
          example: "E"
        legal:
          type: object
          description: Information about tournament legality of this card
          properties:
            standard:
              type: boolean
              description: Indicates whether this card is legal in standard format tournaments
              example: false
            expanded:
              type: boolean
              description: Indicates whether this card is legal in expanded format tournaments
              example: true
        description:
          type: [string, null]
          description: Descriptive text or flavor text on the card
        level:
          type: [number, string]
          description: Level of the Pokemon (can be a string 'X' for LEVEL-UP cards)
          example: 42
        abilities:
          type: array
          description: Pokemon abilities (Poké-Power, Poké-Body, Ability, etc.)
          items:
            type: object
            properties:
              type:
                type: string
                description: Type of ability (Poké-Power, Poké-Body, Ability, etc.)
                example: "Poké-Power"
              name:
                type: string
                description: Name of the ability
                example: "Damage Swap"
              effect:
                type: [string, null]
                description: Effect text of the ability
                example: "As often as you like..."
        attacks:
          type: array
          description: Pokemon attacks
          items:
            required:
              - name
            type: object
            properties:
              cost:
                type: array
                description: Energy cost to use this attack
                items:
                  type: string
                  example: "Psychic"
              name:
                type: string
                description: Name of the attack
                example: "Confuse Ray"
              effect:
                type: [string, null]
                description: Effect text of the attack
                example: "Flip a coin. If heads, ..."
              damage:
                type: [number, string]
                description: Base damage of the attack (can be string for special damage)
                example: 30
        retreat:
          type: [number, null]
          description: Retreat cost of the Pokemon
          example: 3
        dexId:
          type: array
          description: Pokedex number(s) of the Pokemon
          xml:
            name: dexIdItem
            wrapped: true
          items:
            type: number
        weaknesses:
          type: array
          description: Types the Pokemon is weak against
          xml:
            name: weakness
            wrapped: true
          items:
            $ref: '#/components/schemas/WeakRes'
        resistances:
          type: array
          description: Types the Pokemon is resistant to
          xml:
            name: resistance
            wrapped: true
          items:
            $ref: '#/components/schemas/WeakRes'
        boosters:
          type: array
          description: Boosters in which this card is available
          items:
            $ref: '#/components/schemas/Booster'
          example:
            - id: "boo_A1-mewtwo"
              name: "Mewtwo"
              logo: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-logo"
              artwork_front: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-front"
              artwork_back: "https://assets.tcgdex.net/en/tcgp/A1/boosters/mewtwo-front"
        updated:
          type: string
          description: Timestamp of when this card data was last updated
          format: date-time
      # xml:
      #   name: Card
...